refactor: use a service to retrieve conversation details
