+++ src/app/page.tsx
@@ -1,4 +1,5 @@
 // Page: /
+
 import { getConversations } from '@/conversations/services/ConversationService';
 import ConversationFeed from '@/conversations/components/ConversationFeed';
